updating naming for conversion pkg
