+++ api/server/handlers/porter_app/conversion/porter_yaml.go
@@ -1,4 +1,4 @@
-package validation
+package conversion
 
 import (
 	"context"
@@ -11,40 +11,68 @@ import (
 	"github.com/porter-dev/porter/internal/telemetry"
 )
 
-type AppDefinition struct {
-	Name             string `json:"name"`
-	PorterYAMLBase64 string `json:"porter_yaml"`
+// PorterStackYAML represents all the possible fields in a Porter YAML file
+// Either Applications or Services/Apps must be defined, depending on if there are multiple apps
+type PorterStackYAML struct {
+	Applications map[string]*Application `yaml:"applications" validate:"required_without=Services Apps"`
+	Name         string                  `yaml:"name" validate:"required_without=Applications"`
+	Version      *string                 `yaml:"version"`
+	Image        *Image                  `yaml:"image"`
+	Build        *Build                  `yaml:"build"`
+	Env          map[string]string       `yaml:"env"`
+	SyncedEnv    []*SyncedEnvSection     `yaml:"synced_env"`
+	Apps         map[string]Service      `yaml:"apps" validate:"required_without=Applications Services"`
+	Services     map[string]Service      `yaml:"services" validate:"required_without=Applications Apps"`
+
+	Release *Service `yaml:"release"`
 }
 
-func getType(name string, service *Service) (porterv1.ServiceType, error) {
-	var serviceType porterv1.ServiceType
+// Application represents a single app in a Porter YAML file
+type Application struct {
+	Name     string             `yaml:"name"`
+	Services map[string]Service `yaml:"services" validate:"required"`
+	Image    *Image             `yaml:"image"`
+	Build    *Build             `yaml:"build"`
+	Env      map[string]string  `yaml:"env"`
 
-	if service.Type != "" {
-		if service.Type == "web" {
-			return porterv1.ServiceType_SERVICE_TYPE_WEB, nil
-		}
-		if service.Type == "worker" {
-			return porterv1.ServiceType_SERVICE_TYPE_WORKER, nil
-		}
-		if service.Type == "job" {
-			return porterv1.ServiceType_SERVICE_TYPE_JOB, nil
-		}
-
-		return serviceType, fmt.Errorf("invalid service type '%s'", service.Type)
-	}
-
-	if strings.Contains(name, "web") {
-		return porterv1.ServiceType_SERVICE_TYPE_WEB, nil
-	}
-
-	if strings.Contains(name, "wkr") {
-		return porterv1.ServiceType_SERVICE_TYPE_WORKER, nil
-	}
-
-	return porterv1.ServiceType_SERVICE_TYPE_JOB, nil
+	Release *Service `yaml:"release"`
 }
 
-func ConvertAppsFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
+// Build represents the build settings for a Porter app
+type Build struct {
+	Context    string   `yaml:"context" validate:"dir"`
+	Method     string   `yaml:"method" validate:"required,oneof=pack docker registry"`
+	Builder    string   `yaml:"builder" validate:"required_if=Method pack"`
+	Buildpacks []string `yaml:"buildpacks"`
+	Dockerfile string   `yaml:"dockerfile" validate:"required_if=Method docker"`
+	Image      string   `yaml:"image" validate:"required_if=Method registry"`
+}
+
+// Service represents a single service in a porter app
+type Service struct {
+	Run    string      `yaml:"run"`
+	Config interface{} `yaml:"config"`
+	Type   string      `yaml:"type" validate:"required, oneof=web worker job"`
+}
+
+type SyncedEnvSection struct {
+	Name    string                `json:"name" yaml:"name"`
+	Version uint                  `json:"version" yaml:"version"`
+	Keys    []SyncedEnvSectionKey `json:"keys" yaml:"keys"`
+}
+
+type SyncedEnvSectionKey struct {
+	Name   string `json:"name" yaml:"name"`
+	Secret bool   `json:"secret" yaml:"secret"`
+}
+
+type Image struct {
+	Repository string `yaml:"repository"`
+	Tag        string `yaml:"tag"`
+}
+
+// AppProtoFromYaml converts a Porter YAML file into a map of PorterApp proto objects to validate on CCP
+func AppProtoFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
 	ctx, span := telemetry.NewSpan(context.Background(), "convert-apps-from-porter-yaml")
 	defer span.End()
 
@@ -55,7 +83,7 @@ func ConvertAppsFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
 	}
 
 	// convert each app in the app group into a valid standalone definition
-	apps, err := getAppsFromApplicationGroup(*porterYaml)
+	apps, err := appsFromApplicationGroup(*porterYaml)
 	if err != nil {
 		return nil, telemetry.Error(ctx, span, err, "error getting apps from application group")
 	}
@@ -76,12 +104,12 @@ func ConvertAppsFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
 
 		services := make(map[string]*porterv1.Service, 0)
 		for name, service := range app.Services {
-			serviceType, err := getType(name, service)
+			serviceType, err := protoEnumFromType(name, service)
 			if err != nil {
 				return nil, telemetry.Error(ctx, span, err, "error getting service type")
 			}
 
-			validService, err := castServiceConfig(service, serviceType)
+			validService, err := serviceProtoFromConfig(service, serviceType)
 			if err != nil {
 				return nil, telemetry.Error(ctx, span, err, "error casting service config")
 			}
@@ -91,7 +119,7 @@ func ConvertAppsFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
 		validApp.Services = services
 
 		if app.Release != nil {
-			release, err := castServiceConfig(app.Release, porterv1.ServiceType_SERVICE_TYPE_JOB)
+			release, err := serviceProtoFromConfig(*app.Release, porterv1.ServiceType_SERVICE_TYPE_JOB)
 			if err != nil {
 				return nil, telemetry.Error(ctx, span, err, "error casting release config")
 			}
@@ -104,10 +132,10 @@ func ConvertAppsFromYaml(file []byte) (map[string]*porterv1.PorterApp, error) {
 	return validatedApps, nil
 }
 
-func castYamltoApp(porterYaml PorterStackYAML) (Application, error) {
+func yamlToApplication(porterYaml PorterStackYAML) (Application, error) {
 	application := Application{}
 
-	var services map[string]*Service
+	var services map[string]Service
 
 	if porterYaml.Services == nil && porterYaml.Apps == nil {
 		return application, fmt.Errorf("no apps or services defined in porter yaml")
@@ -136,7 +164,35 @@ func castYamltoApp(porterYaml PorterStackYAML) (Application, error) {
 	return application, nil
 }
 
-func castServiceConfig(service *Service, serviceType porterv1.ServiceType) (*porterv1.Service, error) {
+func protoEnumFromType(name string, service Service) (porterv1.ServiceType, error) {
+	var serviceType porterv1.ServiceType
+
+	if service.Type != "" {
+		if service.Type == "web" {
+			return porterv1.ServiceType_SERVICE_TYPE_WEB, nil
+		}
+		if service.Type == "worker" {
+			return porterv1.ServiceType_SERVICE_TYPE_WORKER, nil
+		}
+		if service.Type == "job" {
+			return porterv1.ServiceType_SERVICE_TYPE_JOB, nil
+		}
+
+		return serviceType, fmt.Errorf("invalid service type '%s'", service.Type)
+	}
+
+	if strings.Contains(name, "web") {
+		return porterv1.ServiceType_SERVICE_TYPE_WEB, nil
+	}
+
+	if strings.Contains(name, "wkr") {
+		return porterv1.ServiceType_SERVICE_TYPE_WORKER, nil
+	}
+
+	return porterv1.ServiceType_SERVICE_TYPE_JOB, nil
+}
+
+func serviceProtoFromConfig(service Service, serviceType porterv1.ServiceType) (*porterv1.Service, error) {
 	configYaml, err := yaml.Marshal(service.Config)
 	if err != nil {
 		return nil, fmt.Errorf("Unable to marshal service config: %w", err)
@@ -195,14 +251,14 @@ func castServiceConfig(service *Service, serviceType porterv1.ServiceType) (*por
 	return validSevice, nil
 }
 
-func getAppsFromApplicationGroup(porterYaml PorterStackYAML) ([]Application, error) {
+func appsFromApplicationGroup(porterYaml PorterStackYAML) ([]Application, error) {
 	ctx, span := telemetry.NewSpan(context.Background(), "extract-apps-from-porter-yaml")
 	defer span.End()
 
 	apps := make([]Application, 0)
 
 	if porterYaml.Applications == nil {
-		app, err := castYamltoApp(porterYaml)
+		app, err := yamlToApplication(porterYaml)
 		if err != nil {
 			return nil, telemetry.Error(ctx, span, err, "error getting single app from porter yaml")
 		}
